djangotutorial/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
Flask==3.0.3
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
narwhals==1.28.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
plotly==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
retrying==1.3.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0